pyproject.toml
[project]
name = "install-release"
version = "0.7.1"
readme = "README.md"
description = "Simple package manager to easily install, update and manage any tool directly from github or gitlab releases"
authors = [
  { name = "Rishang", email = "rishangbhavsarcs@gmail.com" }
]

classifiers = [
    "Intended Audience :: Information Technology",
    "Intended Audience :: System Administrators",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Software Development :: Libraries",
    "Topic :: Software Development :: Build Tools",
    "Topic :: Software Development",
    "Typing :: Typed",
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",

]

requires-python = ">=3.10"
dependencies = [
  "python-magic>=0.4.27",
  "pyyaml>=6.0.3",
  "requests>=2.32.5",
  "rich>=14.1.0",
  "typer>=0.17.3",
  "zstandard>=0.24.0",
]

[project.urls]
Homepage = "https://github.com/Rishang/install-release"
Documentation = "https://github.com/Rishang/install-release"
Repository = "https://github.com/Rishang/install-release"
Issues = "https://github.com/Rishang/install-release/issues"
Changelog = "https://github.com/Rishang/install-release/release"


[dependency-groups]
dev = [
  "mypy>=1.17.1",
  "nuitka>=2.7.13",
  "pre-commit>=4.3.0",
  "ruff>=0.12.12",
  "setuptools",
  "types-requests>=2.32.4.20250809",
]

[build-system]
requires = ["setuptools>=42"]
build-backend = "setuptools.build_meta"

[tool.uv]
package = true

[tool.setuptools.packages.find]
include = ["InstallRelease*"]

[project.scripts]
install-release = "InstallRelease.cli:app"
ir = "InstallRelease.cli:app"
